+++ planet-mitchell/src/Optic/Getting.hs
@@ -2,10 +2,6 @@ module Optic.Getting
   ( -- * Getting
     Getting
   , (^.)
-  , view
-  , views
-  , use
-  , uses
   ) where
 
-import Control.Lens.Getter
+import Control.Lens.Getter (Getting, (^.))