+++ planet-mitchell/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to the [Haskell Package Versioning Policy](https://pvp.haskell.org/).
 
+## Unreleased
+
+### Added
+- Export `Mitchell.Prelude.Type`
+- Export `Monoid.mappend`
+- Export `(State..=)`
+- Export `(State.%=)`
+
+### Changed
+- Move `Optic.Getting.view(s)` to `Reader.view(s)`
+- Move `Optic.Getting.use(s)` to `State.view(s)`
+
 ## [0.1.0] - 2018-10-21
 
 ### Added